requirements.txt
fastapi
pydantic<2.0
uvicorn
psycopg2-binary
alembic
python-dotenv
fastapi-crudrouter
